Uninstall scripts when hardlinks are used (#146)

+++ rye/src/installer.rs
@@ -8,6 +8,7 @@ use console::style;
 use once_cell::sync::Lazy;
 use pep508_rs::{Requirement, VersionOrUrl};
 use regex::Regex;
+use same_file::is_same_file;
 use url::Url;
 
 use crate::bootstrap::{ensure_self_venv, fetch};
@@ -107,9 +108,13 @@ pub fn install(
         }
         cmd.env("PYTHONWARNINGS", "ignore");
     }
-    cmd.arg("--")
-        .arg(&requirement.to_string())
-        .arg("importlib-metadata==6.6.0; python_version==\"3.7\"");
+    cmd.arg("--").arg(&requirement.to_string());
+
+    // we don't support versions below 3.7, but for 3.7 we need importlib-metadata
+    // to be installed
+    if py_ver.major == 3 && py_ver.minor == 7 {
+        cmd.arg("importlib-metadata==6.6.0");
+    }
 
     let status = cmd.status()?;
     if !status.success() {
@@ -256,20 +261,22 @@ pub fn uninstall(package: &str, output: CommandOutput) -> Result<(), Error> {
 }
 
 fn uninstall_helper(target_venv_path: &Path, shim_dir: &Path) -> Result<(), Error> {
-    fs::remove_dir_all(target_venv_path).ok();
+    let target_venv_bin_path = target_venv_path.join(VENV_BIN);
+    if !target_venv_bin_path.is_dir() {
+        return Ok(());
+    }
 
-    for script in fs::read_dir(shim_dir)? {
+    for script in fs::read_dir(target_venv_bin_path)? {
         let script = script?;
-        // XXX: this does not handle hardlinks on windows :(
-        if !script.path().is_symlink() {
-            continue;
-        }
-        if let Ok(target) = fs::read_link(&script.path()) {
-            if target.strip_prefix(target_venv_path).is_ok() {
-                fs::remove_file(&script.path())?;
+        if let Some(base_name) = script.path().file_name() {
+            let shim_path = shim_dir.join(base_name);
+            if let Ok(true) = is_same_file(&shim_path, &script.path()) {
+                fs::remove_file(&shim_path).ok();
             }
         }
     }
+
+    fs::remove_dir_all(target_venv_path).ok();
 
     Ok(())
 }
